registry: move to canonical location
Internal libraries live in internal/
