+++ internal/registry/registry.go
@@ -1,4 +1,4 @@
-package main
+package registry
 
 import (
 	"encoding/json"
